backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0
